backend/requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
PyJWT==2.9.0
sqlparse==0.5.3
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg==3.1.18
django-environ==0.11.2
requests==2.32.3
gunicorn==23.0.0
